requirements.txt
karton-core>=4.2.0,<5.0.0
lief==0.11.5
yara-python==4.1.0
